+++ Data/SBV/Control/Utils.hs
@@ -139,8 +139,9 @@ syncUpSolver afterAPush is = do
                        cnsts <- sortBy cmp . map swapc . Map.toList <$> readIORef (rNewConsts is)
                        arrs  <- IMap.toAscList <$> readIORef (rNewArrs is)
                        tbls  <- map arrange . sortBy cmp . map swap . Map.toList <$> readIORef (rNewTbls is)
+                       uis   <- Map.toAscList <$> readIORef (rNewUIs is)
                        as    <- readIORef (rNewAsgns is)
-                       return $ toIncSMTLib afterAPush cfg inps ks cnsts arrs tbls as cfg
+                       return $ toIncSMTLib afterAPush cfg inps ks cnsts arrs tbls uis as cfg
         mapM_ (send True) $ mergeSExpr ls
 
 -- | Retrieve the query context